requirements.txt

# Analysis Requirements

pandas
matplotlib
seaborn
h5py
pooch
torch
scikit-learn
mne


# Docs requirements
sphinx
#sphinx==3.1.1
sphinx-gallery==0.8.1
sphinx_rtd_theme #==0.5.0
sphinx-tabs==1.3.0
sphinx-copybutton==0.3.1
sphinxcontrib-httpdomain==1.7.0
numpydoc==1.1.0
recommonmark==0.6.0
versioneer==0.19
rst2pdf==0.98
pyyaml
nbsphinx
jupyter
ipykernel
jupyter-alabaster-theme
gdown
jinja2<3.1


# Test requirements

mypy
pytest
pytest-cov
nbval


# Types

types-requests



